+++ lets/redirect/tcp/tcp.py
@@ -7,7 +7,7 @@ class TCP(Module):
     Redirect TCP traffic over TCP.
     """
     images = [
-        "local/socat",      # Cross-platform
+        "local/socat:1.0.0",      # Cross-platform
         ]
 
     @classmethod
@@ -26,7 +26,7 @@ class TCP(Module):
         cmd += " tcp-connect:%s:%i" % (rhost, rport)
 
         # Launch redirector
-        with Container.run("local/socat",
+        with Container.run("local/socat:1.0.0",
             network="host",     # Use host network to allow local addresses
             tty=True,
             stdin_open=True,